+++ src/elements/colour.c
@@ -6,7 +6,7 @@
 /*   By: pandalaf <pandalaf@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/15 16:41:50 by pandalaf          #+#    #+#             */
-/*   Updated: 2023/01/24 17:03:20 by pandalaf         ###   ########.fr       */
+/*   Updated: 2023/01/24 19:00:01 by pandalaf         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ t_colour	colour_ambient(t_colour colour, t_ambient *ambient)
 }
 
 //Function works out the lighting effect of a diffuse light on a point. Linear.
-t_colour	colour_diffuse_lin(t_diffuse *difflight, t_point *point)
+t_colour	colour_diffuse_linear(t_diffuse *difflight, t_point *point)
 {
 	t_colour	colour;
 	double		distance;
@@ -76,23 +76,6 @@ t_colour	colour_diffuse_inverse_square(t_diffuse *difflight, t_point *point)
 		colour = WHITE;
 	if (colour < BLACK)
 		colour = BLACK;
-	return (colour);
-}
-
-//Function works out the lighting of an intersection based on elements and lights
-t_colour	colour_lighting(t_objlist *objlist, t_intersect *intersect)
-{
-	t_colour	colour;
-	t_colour	objcolour;
-	t_ambient	*ambient;
-	t_diffuse	*diffuse;
-
-	ambient = ambient_objlist(objlist);
-	diffuse = diffuse_objlist(objlist);
-	objcolour = colour_object(intersect->object);
-	colour = colour_ambient(objcolour, ambient);
-	// colour += colour_diffuse_lin();
-	colour += colour_diffuse_inverse_square(diffuse, intersect->point);
 	return (colour);
 }
